pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-starter-python"
version = "1.0.0"
description = "Simple voice AI assistant built with LiveKit Agents for Python"
requires-python = ">=3.10, <3.15"

dependencies = [
    "empy==3.3.4",
    "httpx>=0.28.1",
    "lark>=1.3.1",
    "livekit-agents[openai,silero,turn-detector]~=1.4",
    "livekit-plugins-noise-cancellation~=0.2",
    "onnxruntime<1.22",
    "openpyxl>=3.1.5",
    "playwright>=1.58.0",
    "PyYAML>=6.0.2",
    "python-dotenv",
    "tavily-python>=0.7.21",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
# Note: addopts is processed after setuptools entry points, so these -p no: flags
# suppress the plugins from causing issues during collection but cannot prevent
# them loading. For Level 2 tests (ROS2 sourced), also set:
#   export PYTEST_ADDOPTS="-p no:launch_ros -p no:launch_testing"
addopts = "-p no:launch_ros -p no:launch_testing -p no:ament_flake8 -p no:ament_lint -p no:ament_copyright -p no:ament_pep257 -p no:ament_xmllint"
pythonpath = ["src"]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "A", "C4", "UP", "SIM", "RUF"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
